+++ tests/integration/helpers_ha.py
@@ -2,7 +2,10 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
+import itertools
 import json
+import secrets
+import string
 import subprocess
 from collections.abc import Callable, Generator
 from contextlib import contextmanager
@@ -12,6 +15,12 @@ import jubilant_backports
 import yaml
 from jubilant_backports import Juju
 from jubilant_backports.statustypes import Status
+from mysql.connector.errors import (
+    DatabaseError,
+    InterfaceError,
+    OperationalError,
+    ProgrammingError,
+)
 from tenacity import (
     Retrying,
     retry,
@@ -22,7 +31,7 @@ from tenacity import (
 
 from constants import ROOT_USERNAME, SERVER_CONFIG_USERNAME
 
-from .helpers import execute_queries_on_unit
+from .connector import MysqlConnector
 
 CHARM_METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
 
@@ -31,6 +40,19 @@ TEST_DATABASE_NAME = "testing"
 
 JujuModelStatusFn = Callable[[Status], bool]
 JujuAppsStatusFn = Callable[[Status, str], bool]
+
+
+def generate_random_string(length: int) -> str:
+    """Generate a random string of the provided length.
+
+    Args:
+        length: the length of the random string to generate
+
+    Returns:
+        A random string comprised of letters and digits
+    """
+    choices = string.ascii_letters + string.digits
+    return "".join([secrets.choice(choices) for i in range(length)])
 
 
 def check_mysql_instances_online(
@@ -599,3 +621,118 @@ def wait_for_unit_message(app_name: str, unit_name: str, unit_message: str) -> J
     return lambda status: (
         status.apps[app_name].units[unit_name].workload_status.message == unit_message
     )
+
+
+async def execute_queries_on_unit(
+    unit_address: str,
+    username: str,
+    password: str,
+    queries: list[str],
+    commit: bool = False,
+    raw: bool = False,
+) -> list:
+    """Execute given MySQL queries on a unit.
+
+    Args:
+        unit_address: The public IP address of the unit to execute the queries on
+        username: The MySQL username
+        password: The MySQL password
+        queries: A list of queries to execute
+        commit: A keyword arg indicating whether there are any writes queries
+        raw: Whether MySQL results are returned as is, rather than converted to Python types.
+
+    Returns:
+        A list of rows that were potentially queried
+    """
+    config = {
+        "user": username,
+        "password": password,
+        "host": unit_address,
+        "raise_on_warnings": False,
+        "raw": raw,
+    }
+
+    with MysqlConnector(config, commit) as cursor:
+        for query in queries:
+            cursor.execute(query)
+        output = list(itertools.chain(*cursor.fetchall()))
+
+    return output
+
+
+def get_read_only_endpoints(relation_data: list) -> set[str]:
+    """Returns the read-only-endpoints from the relation data.
+
+    Args:
+        relation_data: The dictionary that contains the info
+    Returns:
+        a set that contains the read-only-endpoints
+    """
+    related_units = relation_data[0]["related-units"]
+    read_only_endpoints = set()
+    for _, relation_data in related_units.items():
+        assert "data" in relation_data
+        data = relation_data["data"]["data"]
+
+        try:
+            j_data = json.loads(data)
+            if "read-only-endpoints" in j_data:
+                read_only_endpoint_field = j_data["read-only-endpoints"]
+                if read_only_endpoint_field.strip() == "":
+                    continue
+                for ep in read_only_endpoint_field.split(","):
+                    read_only_endpoints.add(ep)
+        except json.JSONDecodeError as e:
+            raise ValueError("Relation data are not valid JSON.") from e
+
+    return read_only_endpoints
+
+
+def get_read_only_endpoint_ips(relation_data: list) -> list[str]:
+    """Returns the read-only-endpoint hostnames from the relation data.
+
+    Args:
+        relation_data: The dictionary that contains the info
+    Returns:
+        a set that contains the read-only-endpoint hostnames
+    """
+    read_only_endpoints = get_read_only_endpoints(relation_data)
+    read_only_endpoint_hostnames = []
+    for read_only_endpoint in read_only_endpoints:
+        if ":" in read_only_endpoint:
+            read_only_endpoint_hostnames.append(read_only_endpoint.split(":")[0])
+        else:
+            raise ValueError("Malformed endpoint")
+    return read_only_endpoint_hostnames
+
+
+@retry(stop=stop_after_attempt(30), wait=wait_fixed(5), reraise=True)
+def is_connection_possible(
+    credentials: dict, *, retry_if_not_possible=False, **extra_opts
+) -> bool:
+    """Test a connection to a MySQL server.
+
+    Args:
+        credentials: A dictionary with the credentials to test
+        retry_if_not_possible: Retry if connection not possible
+        extra_opts: extra options for mysql connection
+    """
+    config = {
+        "user": credentials["username"],
+        "password": credentials["password"],
+        "host": credentials["host"],
+        "raise_on_warnings": False,
+        "connection_timeout": 10,
+        **extra_opts,
+    }
+
+    try:
+        with MysqlConnector(config) as cursor:
+            cursor.execute("SELECT 1")
+            return cursor.fetchone()[0] == 1
+    except (DatabaseError, InterfaceError, OperationalError, ProgrammingError):
+        # Errors raised when the connection is not possible
+        if retry_if_not_possible:
+            # Retry
+            raise
+        return False
