+++ tests/integration/integration/high_availability/test_replication_scaling.py
@@ -7,8 +7,8 @@ import jubilant_backports
 import pytest
 from jubilant_backports import Juju
 
-from ...helpers import generate_random_string
 from ...helpers_ha import (
+    generate_random_string,
     get_app_units,
     insert_mysql_test_data,
     remove_mysql_test_data,
