+++ tests/integration/integration/high_availability/test_self_healing_network_cut.py
@@ -17,16 +17,14 @@ from tenacity import (
 
 from constants import CLUSTER_ADMIN_USERNAME
 
-from ...helpers import (
-    generate_random_string,
-    is_connection_possible,
-)
 from ...helpers_ha import (
     check_mysql_units_writes_increment,
+    generate_random_string,
     get_app_units,
     get_mysql_primary_unit,
     get_unit_ip,
     insert_mysql_test_data,
+    is_connection_possible,
     remove_mysql_test_data,
     verify_mysql_test_data,
     wait_for_apps_status,
