+++ tests/integration/integration/high_availability/test_self_healing_process_frozen.py
@@ -9,16 +9,14 @@ from jubilant_backports import Juju
 
 from constants import CLUSTER_ADMIN_USERNAME
 
-from ...helpers import (
-    generate_random_string,
-    is_connection_possible,
-)
 from ...helpers_ha import (
     check_mysql_units_writes_increment,
+    generate_random_string,
     get_mysql_primary_unit,
     get_unit_ip,
     get_unit_process_id,
     insert_mysql_test_data,
+    is_connection_possible,
     remove_mysql_test_data,
     update_interval,
     verify_mysql_test_data,
