+++ tests/integration/integration/high_availability/test_self_healing_process_killed.py
@@ -7,9 +7,9 @@ import jubilant_backports
 import pytest
 from jubilant_backports import Juju
 
-from ...helpers import generate_random_string
 from ...helpers_ha import (
     check_mysql_units_writes_increment,
+    generate_random_string,
     get_mysql_primary_unit,
     get_unit_process_id,
     insert_mysql_test_data,
