+++ tests/integration/integration/high_availability/test_self_healing_stop_all.py
@@ -9,15 +9,13 @@ from jubilant_backports import Juju
 
 from constants import CLUSTER_ADMIN_USERNAME
 
-from ...helpers import (
-    generate_random_string,
-    is_connection_possible,
-)
 from ...helpers_ha import (
     check_mysql_units_writes_increment,
+    generate_random_string,
     get_app_units,
     get_unit_ip,
     insert_mysql_test_data,
+    is_connection_possible,
     remove_mysql_test_data,
     start_mysql_process_gracefully,
     stop_mysql_process_gracefully,
