+++ tests/integration/integration/roles/test_instance_roles.py
@@ -9,9 +9,9 @@ import pytest
 from jubilant_backports import Juju
 from mysql.connector.errors import ProgrammingError
 
-from ...helpers import execute_queries_on_unit
 from ...helpers_ha import (
     MINUTE_SECS,
+    execute_queries_on_unit,
     get_app_units,
     get_mysql_primary_unit,
     get_mysql_server_credentials,
